Fix AMP float16 overflow: force float32 for softmax, CE, distance
Root cause of periodic loss explosions: under AMP autocast, softmax on
part_logits overflows when logit values exceed ~11 (float16 max = 65504).
Also fix float16 precision issues in euclidean distance and push loss.

- pams.py: F.softmax(part_logits.float(), dim=1)
- make_loss.py: F.cross_entropy(bpa_logits.float(), targets)
- part_loss.py: euclidean_dist and push_loss in float32

+++ loss/make_loss.py
@@ -97,7 +97,7 @@ def make_loss(cfg, num_classes):    # modified by gu
 
                 # BPA supervision (only when available, i.e. training with pose)
                 if extras and 'bpa_logits' in extras:
-                    bpa_loss = F.cross_entropy(extras['bpa_logits'], extras['bpa_targets'])
+                    bpa_loss = F.cross_entropy(extras['bpa_logits'].float(), extras['bpa_targets'])
                     total = total + pams_bpa_w * bpa_loss
 
                 # Push diversity loss
